+++ jobbergate-api/jobbergate_api/apps/job_submissions/constants.py
@@ -17,6 +17,7 @@ class JobSubmissionStatus(str, Enum):
     FAILED = "FAILED"
     UNKNOWN = "UNKNOWN"
     REJECTED = "REJECTED"
+    CANCELLED = "CANCELLED"
 
     @classmethod
     def pretty_list(cls):